backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.1

